ndb - test for creating many hashmaps
the test tries to create as many new (default) hashmaps
as the loop command line option specifies.

if it fail due to error 707/712 (NoMoreTable/HashmapRecords)
the test will succeed anyway.

NOTE: since there is no way to drop hashmaps, running this test
can make it impossible to add nodes or use non default partitioning
since no new hashmap can not be created. 

So this is not added to regular autotest now, and it should definitly
not be runned in production systems!