Bug#16368875 INNODB: FAILING ASSERTION:              PRIMARY_KEY_NO == -1 || PRIMARY_KEY_NO == 0
Problem:
--------
We hit following innodb assert when creating index with
prefix on a TINYBLOB field type.

InnoDB: Assertion failure in thread 5732 in file 
        ha_innodb.cc line 9473
InnoDB: Failing assertion: primary_key_no == -1 || 
        primary_key_no == 0

Analysis:
=========
Suppose we have a innodb table without primary key defined
on it. And if we add new UNIQUE index on a NOT NULL field
without using prefix, then both SQL Layer and Innodb promotes
it as primary key.  Following cases describe current behavior,
Eg. Case 1:
-> CREATE TABLE t1 (f1 CHAR(200));
-> ALTER TABLE t1 ADD UNIQUE INDEX (f1);

Case 2:
-> CREATE TABLE t1 (f1 CHAR(200));
-> ALTER TABLE t1 ADD UNIQUE INDEX (f1(195));
This index is not used as primary key, because we have
partial prefix (195)

Case 3:
-> CREATE TABLE t1 (f1 CHAR(200));
-> ALTER TABLE t1 ADD UNIQUE INDEX (f1(200));
The above index is used as primary key, because we have
full prefix (f1(200)) and is equivalent to just (f1).
MySQL just drops the prefix, SHOW CREATE will not show prefix.

The above behavior is applicable to BLOB fields too.
Following cases are for TINYBLOB,
Case 4:
-> CREATE TABLE t1 (f1 TINYBLOB);
-> ALTER TABLE t1 ADD UNIQUE INDEX (f1(200));
This index is not used as primary key, because we have
partial prefix (200)

Now, following is a special case which is buggy,
Case 5:
-> CREATE TABLE t1 (f1 TINYBLOB);
-> ALTER TABLE t1 ADD UNIQUE INDEX (f1(255));
Similar to Case 3, we expect f1 to be primary.
The bug is that SQL layer fail to consider this as primary,
and innodb add's it as primary key. This is root cause of
the assert.

There is also another bug,
Case 6:
-> CREATE TABLE t1 (f1 TINYBLOB);
-> ALTER TABLE t1 ADD UNIQUE INDEX (f1(256));
The above index is added successfully without reporting
an error, when prefix is greater than TINYBLOB max size
i.e., 255.

Note:
* We cannot reproduce problem like in Case 5 using other BLOB
types are the max prefix for these types is greater than
the prefix supported by engine.
* Note that SHOW CREATE TABLE retains the prefix for BLOB
fields even when prefix size is equal to BLOB field max size.
This is required to enable backup and restore schema. This is
not the case with CHAR or VARCHAR.
* The fix is applicable to TEXT field types also.

Fix:
===
We fix Case 5 and Case 6 scenarios.

Fixed function open_binary_frm() to consider BLOB fields
as primary, when the prefix size is equal to the size of BLOB
field max size.

Fixed function mysql_prepare_create_table() to consider
key column size also, when we check prefix size.