BUG#16460123- NOTE ABOUT "MASTER.INFO REPOSITORY" ON CHANGE MASTER TO IS MISLEADING BUG#16461303- WARNING ISSUED ABOUT MASTER.INFO WHEN USING MASTER-INFO-REPOSITORY=TABLE
Problem:
When executing CHANGE MASTER TO on MySQL 5.6 with either
MASTER_USER or MASTER_PASSWORD options, MySQL issues a 
warning:
Level: Note
Code: 1760
Message: Storing MySQL user name or password information
in the master.info repository is not secure and is 
therefore not recommended. Please see the MySQL Manual for
more about this issue and possible alternatives.

This appears to suggest that "master.info" is insecure,
which may lead the reader to believe that they should use
master-info-repository=table instead of 
master-info-repository=file. However, this warning is
issued when using both types of master info repository.

Also, the suggestion to see the MySQL Manual is not
pointing to any specific section in the Manual for user's
convenience.

Fix:
This Note is changed to specify that master info repository
is insecure to store USERNAME or PASSWORD and thence points
to the START SLAVE section of MYSQL Manual for reference.