Bug#15963619 GETTING "TABLE IS READ ONLY" ERROR EVEN WHILE READING WITH --INNODB-READ-ONLY
rb://2439 Approved by Jimmy

This fix is based on the Sunny's patch in the bug report.

When InnoDB is in read-only mode, we prevent all lock operation in
ha_innodbase::external_lock as MySQL want to create table.
In fact, we only need to prevent write lock operation when MySQL is
doing create table.