Fix for Bug#16804581 ASSERTION FAILED: OTHER_TBLS_OK || IDX_COND->CONST_ITEM()
The failing asserts happens when considering pushing an index
condition and the condition does not contain fields from the index. In
this is the case we do not push the index condition to the storage
engine. The failing assert is to detect that this only happens when
the condition can access fields from earlier tables in the join or is
a constant item. In the failing query, the index condition is an
updatable user variable which caused the assert to fail.

The fix for this is to extend the assert to also allow conditions
containing updatable user variables to be rejected as index pushdown
conditions if the condition does not contains any fields from the
index.