Bug#16875543 MEMCACHED: SERVER CRASH IN MEMCACHED SET OPERATION.
rb://2533 Approved by Jimmy

This bug is because, we use dtype len to pad the utf8 char(n) column
value.
For example, there's a utf8 char column c1 char(120). The dtype len of
this column is 3*120=360.
In memcached update operation, we copied a field from an old tuple, and
the actual data length is less than 360. Since we optimize the utf8
string storage for saving space.
The solution is we don't copy old tuple to new tuple anymore. Every time
we create a new tuple, we allocate the memory for it and set it.