BUG#15880018 - MISLEADING MESSAGES AT STARTUP TIME
DESCRIPTION:

When the MySQL server starts, it checks for the file containing
list of all the MySQL error messages. This file by default is
'errmsg.sys'. The location of the error-message file can be
specified by the server option '--lc-messages-dir'. If MySQL
server can not find the file at the location specified by the
user, an error message is displayed and server does not start.
The Error message is general and does not point user in the right
direction.

FIX :

Changed the error message so that user can check
'--lc-messages-dir' configuration directive too if MySQL server
can not find the 'errmsg.sys' file.