This commit fixes bug#16853897 'NDB_JOIN_PUSHDOWN_BKA.TEST FAILS ON SOL10-SPARC'. This bug caused queries using BKA to give wrong results in certain cases. This happens if all of the following are true: - mysqld runs on a big-endian machine. - the table that is buffered in the BKA join cache use a storage   engine with little-endian data format (e.g. InnoDB or MyISAM). - The next table use a storage engine with native-endian data format   (i.e. ndb).
The keys used for accessing the native-endian table is then in little-endian
format, giving wrong results. This is caused by an optimization in the BKA
code: when possible, it will use a pointer into the BKA chache as a key, 
instead of running the regular code for extracting fields from a record
and building a key. There is a method called 
JOIN_CACHE_BKA::check_emb_key_usage() that checks if this optimization is
possible. The fix consists in adding an extra test to this method, to check
if both tables have the same endianness.

Presumably the error would also happen if the two tables were switched, such
that native-endian records were buffered in the BKA cahce and used for building
keys for a little endian table. But I was not able to write a query that would
use BKA this way.

This bug breaks an exiting ndb regression test. Unfortunately there does
not seem to be any way to reproduce this error without ndb, since this
is the only native-endian storage engine.