Bug #16593604 START BACKUP WAIT STARTED WAITS FOREVER ON BACKUP ERRORS
While aborting a backup which is run as WAIT STARTED or WAIT COMPLETED, 
the master data node needs to send out two signals to the mgmt server: 
one to abort the backup and one to log a BackupAborted event. However, 
the backup abort signal is sent only if the backup is started with WAIT 
COMPLETED. For WAIT STARTED, the mgmt server keeps waiting for a backup 
completion/abort signal from the ndbd which is never sent. The ndb_mgm 
client hangs waiting for a result from the mgmt server.

This was fixed by modifying the check for sending the backup abort signal 
so that the backup abort signal will be sent for backups with WAIT 
STARTED as well as backups with WAIT COMPLETED.