BUG#11754116: COLUMN DEFAULT VALUE NOT REFLECTED IN SHOW CREATE               OUTPUT
Analysis:
--------
New default value set for a column using
ALTER TABLE:ALTER COLUMN:SET DEFAULT command is not reflected:
a) If the column is of type TIMESTAMP or DATETIME and
b) Has CURRENT_TIMESTAMP as its default value when it was created.

In this case, when the default value is altered for the
TIMESTAMP/DATETIME column, the default value for the
TIMESTAMP/DATETIME is updated correctly in the table
definition. But since the 'unireg_check" was set to
"TIMESTAMP_DN_FIELD/TIMESTAMP_DNUN_FIELD", while inserting
a tuple, the default value for the field uses 'NOW()' instead of
reading the default value from the table definition.
While displaying table definition also, 'CURRENT_TIMESTAMP' is
displayed because of the 'unireg_check' value.

This means that the bug affected not only SHOW CREATE, but
also could cause wrong values to be inserted in a TIMESTAMP/DATETIME
column if the default value was changed.

Fix:
---
In the current code, flag for unireg_check was always set to
'TIMESTAMP_DN_FIELD/TIMESTAMP_DNUN_FIELD' when current
default value for field is CURRENT_TIMESTAMP. Value
of the 'unireg_check' was not getting updated while
changing the default value of TIMESTAMP/DATETIME field.
As a fix for this patch, added a logic to set the
unireg_check field to proper value while updating the
default value.