Bug#16945618 PERFORMANCE SCHEMA, DEGRADATION UNDER STRESS WHEN INTERNAL BUFFERS ARE FULL
The performance schema uses internal buffers of a fixed size.
The size comes from configuration options.

When the sizing is adequate for the workload, things work properly.

When the sizing is too tight for the workload, the performance schema
spends a lot of time in an internal spin loop to attempt to allocate a memory
buffer, and fails.

Before this bug, this spin loop would happen for each call to instrument an
object, which can be a very frequent operation.

With this fix, when a failure to allocate memory is noticed, a flag is set 
that indicates the corresponding buffer is full.

Subsequent call to allocate memory from the same buffer will fail 
immediately, avoiding the need for a spin loop that fill fail anyway.

When an object is released, the corresponding buffer is not flagged as full
any more, indicating that memory allocation can be attempted again.

With this fix, when running the performance schema with undersized buffers,
the functional result is unchanged: some events are lost and reported as 
such. What is changed is that the performance degradation in this mode, 
which could be catastrophic in some cases, is now avoided.