Bug#16078943: ZERO DATE CAN BE INSERTED IN STRICT NO-ZERO MODE               THROUGH A DEFAULT VALUE.
Analysis:
-------------
The problem was that the default values for a table were
checked for consistency only at CREATE TABLE time. The
consistency checks very much depend on the sql_mode value.
The default values were not checked at the INSERT/UPDATE
time. Thus, if sql_mode became strict in the middle between
CREATE TABLE and INSERT/UPDATE, inconsistent data can be put
into the table.

So here, when sql_mode='NO_ZERO_DATE,NO_ZERO_IN_DATE,
STRICT_ALL_TABLE'(strict mode) then creating table with
default "value 0" or default "value having 0 in date" for
field types DATE/DATETIME/TIMESTAMP" is not allowed. But when
sql_mode is "NO_ZERO_IN_DATE,NO_ZERO_DATE" or "" (non-strict mode),
tables can be created.

If table with default "value 0" or default "value having 0 in
date" for field DATE/DATETIME/TIMESTAMP is created in non-strict
mode and tuple for the table is inserted with default for
DATE/DATETIME/TIMESTAMP fields in strict mode then no error or
warning is generated.

While inserting/updating tuple(s), we are not validating the
default values. While creating the table the default value set 
might be correct but same values might not be correct while
inserting/updating the tuple (one of the reason is sql_mode).

Fix:
-------------
Added a logic to verify the default values while inserting
or updating a tuple. 

Currently the TIMESTAMP data type differs in nonstandard
ways from other data types. But this behavior can be
turned off using "explicit_defaults_for_timestamp" system
variable at server startup. With this patch, old behavior
of TIMESTAMP data type is not changed. But with 
"explicit_defaults_for_timestamp", logic is added to 
verify the default value while inserting or updating
a tuple.

With this, behavior of insert or update with default values
for DATE/DATETIME/TIMESTAMP is as below,

+---------------+------------+----------------+-------------+
|               |            |                |             |
|TYPE/sql_mode  | ""         |NO_ZERO_DATE/   |strict_mode  |
|        value  |            |NO_ZERO_IN_DATE |             |
|---------------+------------+----------------+-------------|
|DATE           | Success    | Warning        | Error       |
|---------------|------------|----------------|-------------|
|DATETIME       | Success    | Warning        | Error       |
|---------------|------------|----------------|-------------|
|TIMESTAMP      |            |                |             |
|+ explicit_-   |            |                |             |
|  defaults_-   |            |                |             |
|  for_-        | Success    | Success        | Success     |
|  timestamp    |            |                |             |
|  = false      |            |                |             |
|(for           |            |                |             |
| backward      |            |                |             |
| compatibility)|            |                |             |
|---------------|------------|----------------|-------------|
|TIMESTAMP      |            |                |             |
|+ explicit_-   |            |                |             |
|  defaults_-   | Success    | Warning        | Error       |
|  for_-        |            |                |             |
|  timestamp    |            |                |             |
|  = true       |            |                |             |
|               |            |                |             |
+-----------------------------------------------------------+


Summary of discussion on impact on Replication:
-----------------------------------------------

*. Found following impact on replication if this issue is 
   fixed,

   Older master (bug is there) --> Newer slave (bug is fixed)

   Let's assume sql_mode is propagated properly between the
   master and slaves. I.e. at any point in time the master 
   and slaves have the same sql_mode value.

   The following sequence is permitted on the master:
     - sql_mode = ''
     - CREATE TABLE t1 (f1 int,
                        f2 DATE DEFAULT 0,
                        f3 DATETIME DEFAULT 0,
                        f4 TIMESTAMP DEFAULT 0);
     - set sql_mode='NO_ZERO_DATE,NO_ZERO_IN_DATE,
                     STRICT_ALL_TABLES';
     - INSERT INTO t1(f1) VALUES (10);

   The result of this sequence on the master is a new record:
      { 0, 0, 0, 0 }
   (that is the bug, because of effective NO_ZERO_DATE &
    NO_ZERO_IN_DATE).

   On the slave, the INSERT statement will fail (because the
   bug is fixed on the slave), thus the replication will stop. 

*. Approaches identified after discussing with Replication team,
   Runtime team and support are,

  1. Introduce a switch to relax the sql_mode during an
     upgrade on the slave. But:

     - This switch is effectively overriding the SQL_MODE from
       the master, thence all sorts of different things can
       go wrong.
     - This switch is only necessary for upgrades, which have
       a solid workarounds: using RBR or relaxing the SQL_MODE
       on the master.
     - Introducing another switch requires double testing and
       is error prone, since users may use it for other cases
       than upgrades and then we will get all sort of weird
       bug reports...
     - This approach does not scale in terms of development
       and maintenance. The use case does not seem to be worth
       it this kind of approach.

  2. Introduce another switch as in suggestion #1, that makes
     the slave behave as the master just for the sole purpose
     of a set of SQL_MODE combinations. Same problems as
     suggestion #1.

  3. Document that during upgrades, the master needs to relax
     the sql_mode or switch to RBR, but note the following
     restrictions:

     - user needs to wait for the old binary logs to be consumed
       before actually be able to upgrade the slaves (so that
       there is no slave still replicating "old" binlogs after
       it is upgraded)
     - user needs to cope with side effects of relaxing the
       sql_mode or use RBR during the upgrade period. 

After considering the risk involved in the approach 1 and 2, its
decided to go with approach 3.

Attached discussion mails in the bug report.