ndb - Improve node restart handling of test programs.
This patch improves handling of node restarts in autotest test programs by
also waiting for Ndb_cluster_connection to be connected when waiting for node
to become started.

This addresses the following not that uncommon scenario:
2-node cluster
restart node 1,
wait for node state at ndb_mgmd to become started
  however, this does not mean that local transporter is connected...
  only that ndb_mgmd is

restart node 2
  ndbapi threads running transactions can then get 4009
  since local transporter never got connected to node 1
    and now node 2 also disconnected.

The patch adds an optional Ndb_cluster_connection argument to NdbRestarter,
that if supplied, will make NdbRestarter also check it for connectivity
when waiting for nodes (or cluster) to become started.

The argument has been added to NdbRestarts, which handles e.g 
NFDuringNodeRestart and MixedReadUpdateScan