Bug#16939689 VALGRIND ERROR IN PFS_SPAWN_THREAD
Before this fix, valgrind errors could be reported under stress tests.

In particular, the following line in pfs_spawn_thread()
  memcpy(pfs->m_username, parent->m_username, sizeof(pfs->m_username));

could cause memcpy to be called for the same source and destination,
which should not happen.

The root cause, by analysis, is that the code is reading attributes 
from the parent thread, while the parent thread instrumentation is already 
destroyed (and was reused for the child, triggering valgrind complaints).

The fix is to capture all the parent attributes needed:
- the thread internal id
- the thread user name
- the thread host name
in the parent thread, before spawning a child,
and use that to initialize the child instrumentation,
instead of pasing a pointer that might become invalid.