Bug#16316564: Crash in do_copy_not_null()
The problem here is a failure in do_copy_not_null() because the pointer
copy->from_null_ptr is NULL, and hence accessing its contents fails.
If we look at the similar function do_copy_null(), it has been enhanced
with a NULL pointer check for the same pointer.
The pointer is NULL for any field that is originally not nullable, but
we may still have a NULL value because of the table->null_row
technique, used for outer-joined tables. The table t2 is semi-joined
to the table t1 (alias z), which is outer-joined, hence it also gets
the outer-join capability, and do_copy_null() has to be used even though
the field is originally not nullable.

I expect the missing NULL pointer check in do_copy_not_null() to be
an oversight, and adding it indeed fixes the problem.

sql/field_conv.cc
  In do_copy_not_null(), add NULL pointer check for copy->from_null_ptr.