Bug#16620047 INCORRECT QUERY RESULT (AFTER SERVER UPGRADE) Regression is introduced in Bug11760517. According to the comment for the Bug11760517 fix, there is an additional cleanup in field.h(Field::is_null() function): Field::is_null() now checks table->null_row before checking the NULL bits.
Order of NULLity checking is changed, firts it checks table->null_row
and then null_ptr[row_offset]. In our case when we evaluate first
result record, evaluate_null_complemented_join_record() function
is called for outer table and it sets table->null_row to true.
Then second result record field uses table->null_row to check
if the field is null and result is NULL instead of real value.

The fix is partial reversion of the field.h change
(return original order of NULL check). Unfortunately this
is not sufficient. A lot of test fails in the main suite
after reversion. It happens because of another additional
cleanup introduced Bug11760517 fix(see the comment):
   @ sql/table.h
     mark_as_null_row() no longer sets the NULL bits of the buffer
      
Reversion of this change fixes all the problems except one, test case
for Bug13541761 which is actually another regression of Bug11760517 fix.
The fix of this problem is to not touch TABLE::null_flags during
JOIN::cleanup.