Bug#16436014: CRASH ITEM::SAVE_IN_FIELD, WITH ROLLUP, DEFAULT
Problem:
If table has been marked as containing only NULL values
for all fields if it is a NULL-complemented row of an OUTER JOIN
then rollup on the field which cannot be nullable results in crash.

Analysis:
In the query presented, as the join results in only NULL rows,
table->null_row is set and as a result, we do not evaluate the result.
But while checking the result of rollup_field, we check the
null value of rollup field. As this was not updated, we try to
access the result and crash.

Fix:
Update the null value of the rollup_field after evaluting the result.