Bug#16916596: "ORDER BY...LIMIT" OPTIMIZATION CHOSE               NON-ORDERING INDEX
Consider a table with a compound index "idx (kp1, kp2, kp3)"
and the query "SELECT ... ORDER BY kp3 LIMIT x".

For this query, the optimizer may decide that although index 'idx'
did not seem like the best way to access the table in the first 
place, it may still be a good index when LIMIT is taken into 
account iff index 'idx' gives correct ordering. The reason is
that even an expensive index (index that produces too many 
rows) like 'idx' may be cheap if execution is allowed to exit
early due to the LIMIT.

However, for index 'idx' to provide sorted rows, the query
must have equality predicates on kp1 and kp2. In this bug, 
the query only had equality predicate on kp1. Even so, the 
ORDER BY... LIMIT optimization was used to switch to index 
'idx', effecively chosing a much worse index without achieving 
the intention.

The fix is to only consider indexes that provide correct 
ordering. This previously worked for single keypart indexes but
not multiple keypart indexes. However, instead of adding yet 
another if() block, all the tests for applicable ORDER BY...
LIMIT indexes in make_join_select() are replaced by a call to
test_if_order_by_key() which exists for the very same 
purpose - to answer the question "does this index provide 
ordered output".