Bug#16970960 NDB_RESTORE_MISC FAILS ON WINDOWS
The Windows error code ERROR_FILE_EXISTS, which is returned from a
file-open operation, is directly copied as the NDB error. This error
code does not correspond to any of the valid NDB errors, therefore it is
treated as an unknown error. The ndb_restore_misc test fails on
receiving this unknown error code instead of the expected FileExists NDB
error.

Added fix to convert the Windows ERROR_FILE_EXISTS error code to the
corresponding NDB FileExists error code.