Bug#16369522: Assertion failed: !tables || thd->lex->is_query_tables_locked()
The server exits when presented with a query like the following

SELECT (SELECT 1 FROM t1) IN (SELECT a FROM t1);

when attempting to evaluate the constant left-hand argument to the
IN subquery predicate. This constant evaluation occurs already in
Item_in_subselect::fix_fields(), through the call convert_constant_item().

Even though the subquery is const, we do not allow execution of subqueries
before tables have been locked.

The problem is the Item_cache object that is used to wrap the left-hand
argument to the IN predicate. It does not propagate the correct const-ness
attributes, and hence the resolver thinks it is safe to evaluate the
subquery. This happens because the Item_cache is not a first class
Item object. Among other things, it does not have a fix_fields()
function. Instead, it relies on ad-hoc code to assign values to
attributes such as used_tables() and const_item().

In this particular use of Item_cache, it also adjusts used_tables
information for the left-hand expression: During subquery transformation,
the left-hand expression and the expressions selected from the subquery
are converted to an equality that is attached to the subquery's query
block. Originally, the LHE is resolved in scope of the outer query block
and thus has used_tables information relative to this, but this
information has to be replaced with an OUTER_REF indicator when being
attached to the inner query block. The Item_cache object took care of
this by explicitly setting the OUTER_REF bit when a table was referenced
in the LHE.

However, when building the equality predicate, an Item_direct_ref
object is always used to wrap the included expressions. This object
can naturally handle the OUTER_REF by setting depended_from for the
object to the outer query block (together with select_lex->context
pointing to the inner query block).

The implementation of Item_ref::const_item() returns FALSE for all
objects that have non-zero used_tables() which is true because the
OUTER_REF indicator is set, and hence early evaluation is prevented.

Notice also that const items need not be designated as outer.
Ignoring those occurrences was required to preserve multiple equality
handling for equalities involving constant values.

We have a change in EXPLAIN for the following type of query:

SELECT col_varchar_key
FROM t1
WHERE ((SELECT i1 FROM t0 WHERE i1 = 7),
       (SELECT i1 FROM t0 WHERE i1 = 2)) NOT IN
    (SELECT col_int_key, col_int_nokey FROM t2);

When transforming the NOT IN subquery using IN-TO-EXISTS, the subqueries in
the left hand side are tested for const-ness. The tests fail because tables are not yet locked at this point. Hence, left->depended_from is set to the outer
query block.

As a result, in get_store_key(), a store_key_item is created instead of a
store_key_const_item, which causes explain_ref_key() to print "func" for
the key reference instread of "const".

A similar query is causing change in
mysql-test/suite/opt_trace/r/bugs_no_prot_none.result.

sql/item.h
  Class Item_cache needed a specific implementation of
  resolved_used_tables() that forwarded the call to the example member.

sql/item_cmpfunc.cc
  In Item_in_optimizer::fix_left(), used tables setting for cache object
  could be simplified.

sql/item_subselect.cc
  In Item_in_subselect::single_value_transformer() and
  Item_in_subselect::row_value_in_to_exists_transformer, when wrapping
  the left-hand expressions in an Item_direct_ref object, set the
  depended_from field of the Item_direct_ref to point to the query block
  where the subquery predicate is located, unless the expression is const.