Bug 16585455 OUT-OF-RANGE BACKUP IDS ARE TRUNCATED SILENTLY
The backup ID which is provided in the START BACKUP command is not 
checked to ensure that it is within the allowed range. Therefore 
out-of-range backup IDs are converted into valid backup IDs by 
overflow instead of causing an error.
      
Added range checking for the backup ID in ndb_mgm for START BACKUP.