Bug#16960800: RANGE OPTIMIZER BELIEVES THAT GEOM FUNCTIONS                     RETURN ROWID ORDERED RESULTS
The range optimizer needs to know if the rows returned from a 
range scan are ordered on ROWID. Multiple indexes that have 
this property can be intersected or union'ed.
      
However, the range optimizer believes that any GEOM function 
on a spatial index returns rows in ROWID order. This is not 
correct. The result is that the range optimizer may choose to 
do index merge intersect or index merge union involving a 
spatial index. This may in turn result in missing rows.

The fix is to let ha_myisam::index_flags() return the 
HA_KEY_SCAN_NOT_ROR bit for spatial indexes.