Bug #16751784 WASTED WORK IN METHOD MYISAMCHK Bug #17001703 WASTED WORK IN METHOD MI_TEST_IF_SORT_REP() Bug #17007268 WASTED WORK IN METHOD RECREATE_TABLE()
Problem: Unwanted execution are happening in the following functions
MYISAMCHK(), MI_TEST_IF_SORT_REP(), RECREATE_TABLE() within a loop
which is consuming some cpu cycles.
      
Fix: For bug#16751784 fix is to break out of loop when variable 
update_index is set to 0. For bug#17001703 fix is to break out 
of loop when variable force is set to 0. For bug#17001703 fix
is to break out of loop when variable "pack" is FALSE.
Fix was contributed by  Po-Chun Chang.