Bug #16971432 MYSQL CLIENT WASTE TIME IN SPRINTF MAKING INFO MESSAGES THAT               ARE NOT PRINTED
Problem: When the client program is in batch mode by default the message
like  "Query OK, 1 row affected" are not shown, but instead the buffer is
prepared using sprintf call which takes some cpu time when the import file
is too large.

Fix: Fix is to check if the client is running in batch mode then skip the
sprintf statement in /client/mysql.cc in com_go(). put_info ignores the 
buffer built by prepare only when in batch mode and verbose < 2. Hence 
skip the sprintf in com_go() only in this case.