Bug #14623333 	MYSQL CLUSTER - NDBREQUIRE CRASH IN EXECGCP_TCFINISHED, DBDIH (LINE: 10381)
There was a DIH-TC race in node failure handling during a 
GCP_COMMIT protocol round.

During node failure, GCP_COMMIT is paused to allow TC takeover on the
DIH/TC Master node to commit/abort the transactions coordinated by the 
failed node.

Where the TC Master has already 'committed' the GCP, it is asked to do
so again, including any taken-over transactions.

There was a race between :
  1.  DIH Master being informed of the node failure, and informing
      TC Master of the need to re-commit the GCP
  2.  TC Master informing DIH-Master of the commit of the GCP 
      (via 'DIH-Slave')

If these two interactions were overlapping then the DIH Master could 
incorrectly assume that the GCP was committed, and potentially proceed
to the next.  This could later result in assertion failures and crashes.

The fix is to be more careful here, and :
  1.  Ensure that only one request from DIH-Master to TC-Master to 
      commit a GCP is in-flight at a time, simplifying the identification
      of responses.
  2.  Include the highest handled node failure number in the TC response,
      so that DIH-Master can detect a race, and send a new commit request.

A new testcase is added to the daily-devel series.